requirements.txt
# Phase 1 - Core Dependencies
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pandas==2.1.3
yfinance==0.2.28
requests==2.31.0
anthropic==0.7.8
python-dotenv==1.0.0
pydantic==2.5.0
supabase==2.3.0
gotrue==2.3.0

# Phase 2 - Advanced Analytics
ta==0.10.2
newsapi-python==0.2.7
matplotlib==3.8.2
plotly==5.17.0

# Phase 3 - Deployment & Monitoring
schedule==1.2.0
jinja2==3.1.2
aiofiles==23.2.1

# Development
pytest==7.4.3
black==23.11.0
flake8==6.1.0